update test file names
